requirements.txt
asgiref==3.8.1
boto3==1.34.141
botocore==1.34.141
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.0.6
django-autoslug==1.9.9
django-crispy-forms==2.2
django-environ==0.11.2
django-mathfilters==1.0.0
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-utils-six==2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
idna==3.7
jmespath==1.0.1
packaging==24.1
phonenumbers==8.13.40
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
